Breakdown Background.js
- First attempt breaking down this file to accomplish:
1. Writing in more documentation via jsDoc
2. Making the code more readable
3. Removing utility code into modules
4. Making the code easier to migrate to potential MV3 port (#17268)
