+++ chromium/background-scripts/modules/extension_state.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * @module Update States
+ * @description Needed updates for disabled, on, and EASE mode extension states
+ */
+
+(function(exports) {
+})(typeof exports == 'undefined' ? require.scopes.state = {} : exports);
