+++ chromium/manifest.json
@@ -9,7 +9,11 @@
     "background": {
         "scripts": [
             "background-scripts/bootstrap.js",
-            "background-scripts/util.js",
+						"background-scripts/modules/log.js",
+						"background-scripts/modules/validation.js",
+						"background-scripts/modules/ip_utils.js",
+						"background-scripts/modules/browser_session.js",
+						"background-scripts/modules/ssl_codes.js",
             "wasm/https_everywhere_lib_wasm.js",
             "background-scripts/wasm.js",
             "background-scripts/update_channels.js",
@@ -18,7 +22,6 @@
             "background-scripts/store.js",
             "external/pako-1.0.5/pako_inflate.min.js",
             "background-scripts/incognito.js",
-            "background-scripts/ip_utils.js",
             "background-scripts/background.js"
         ]
     },
